refactor(internal): move internal directories into trame.internal
This moves most directories reserved for internal use into trame.internal.
