+++ setup.cfg
@@ -31,4 +31,4 @@ include_package_data = True
 
 [options.entry_points]
 console_scripts =
-    trame = trame.dev:main
+    trame = trame.internal.dev:main
