+++ trame/html/deckgl/__init__.py
@@ -1,4 +1,4 @@
-from trame.app import get_app_instance
+from trame.internal.app import get_app_instance
 from trame.html import AbstractElement
 from pywebvue.modules import Deck
 import json
