+++ trame/html/vtk/common.py
@@ -1,4 +1,4 @@
-from trame.app import get_app_instance
+from trame.internal.app import get_app_instance
 from trame.html import AbstractElement
 
 MODULE = None
