+++ trame/internal/utils/cli.py
@@ -8,7 +8,7 @@ def get_cli_parser():
     >>> args, unknown = parser.parse_known_args()
     >>> print(args.output)
     """
-    from trame.app import get_app_instance
+    from trame.internal.app import get_app_instance
 
     _app = get_app_instance()
     return _app.cli_parser
