+++ trame/internal/utils/desktop.py
