+++ trame/internal/utils/string.py
